moved UI toolkit code to own files
