+++ Assets/Mirage/Editor/ReadOnlyDecoratorDrawer.cs
@@ -1,18 +1,10 @@
-﻿#if UNITY_2022_2_OR_NEWER
-#define USE_UI_TOOLKIT
-#endif // UNITY_2022_2_OR_NEWER
-
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
-#if USE_UI_TOOLKIT
-using UnityEditor.UIElements;
-using UnityEngine.UIElements;
-#endif // USE_UI_TOOLKIT
 
 namespace Mirage
 {
     [CustomPropertyDrawer(typeof(ReadOnlyInspectorAttribute))]
-    public class ReadOnlyDecoratorDrawer : PropertyDrawer
+    public partial class ReadOnlyDecoratorDrawer : PropertyDrawer
     {
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
@@ -25,14 +17,5 @@ namespace Mirage
         {
             return EditorGUI.GetPropertyHeight(property, label, true);
         }
-
-#if USE_UI_TOOLKIT
-        public override VisualElement CreatePropertyGUI(SerializedProperty property)
-        {
-            var field = new PropertyField(property);
-            field.SetEnabled(false);
-            return field;
-        }
-#endif // USE_UI_TOOLKIT
     }
 }
